split logic into before and after query
So the same eval(pm.globals.get('magic'))
can be used both in scripts
after response and before query.
