+++ src/index.ts
@@ -1,2 +1,2 @@
-import "./guard";
-import "./magic";
+import "./guard_before_query";
+import "./test_after_response";
