+++ src/test_after_response.ts
@@ -1,36 +1,11 @@
-'use strict';
-
-function guard_before_query() {
-  const env_name_like = magic.guard?.env_name_like;
-  const curr = pm.environment.name;
-  let ok = env_name_like ? false : true;
-  if (env_name_like) {
-    if (Array.isArray(env_name_like)) {
-      ok = env_name_like.some((like) => new RegExp(like).test(curr));
-    } else {
-      ok = new RegExp(env_name_like).test(curr);
-    }
-  }
-  if (!ok) {
-    throw new Error(`Use with "${env_name_like}"-like named environment!`);
-  }
-}
-if (pm.info.eventName === "beforeQuery") {
-  guard_before_query();
-}
-
-function mapping(mapping2, source, destination, prefix = "") {
-  if (!mapping2) return;
-  Object.entries(mapping2).forEach(([k, _path]) => {
-    const value = _path.reduce((acc, p) => acc[p], source);
-    console.debug(`Set ${destination}.${prefix + k} = ${value}`);
-    value !== void 0 && pm[destination].set(prefix + k, value);
-  });
-}
+import { mapping } from "./lib/mapping";
+import * as _types from "./types";
+_types;
 
 if (pm.info.eventName === "afterResponse") {
   test_after_response();
 }
+
 function test_after_response() {
   pm.test(pm.request.name, () => {
     const {
@@ -44,9 +19,10 @@ function test_after_response() {
       req_fbody_to_env,
       req_fbody_to_globals,
       req_fbody_to_col,
-      prefix
+      prefix,
     } = magic;
-    const actual_res_code = pm.response.code || pm.response.transport.http?.statusCode;
+    const actual_res_code = pm.response.code ||
+      pm.response.transport.http?.statusCode;
     if (!actual_res_code) {
       console.warn("Unable to get res status code... Please open the issue");
     } else if (!res_codes.includes(actual_res_code)) {
@@ -54,7 +30,9 @@ function test_after_response() {
       pm.expect(res_codes).includes(actual_res_code);
       return;
     }
+
     let jData = {};
+
     try {
       jData = pm.response.json();
     } catch (err) {
@@ -64,8 +42,8 @@ function test_after_response() {
     let fData = {};
     try {
       fData = pm.request.body.formdata.toObject();
-    } catch {
-    }
+    } catch {}
+
     mapping(res_jbody_to_env, jData, "environment", prefix);
     mapping(res_jbody_to_col, jData, "collectionVariables", prefix);
     mapping(res_jbody_to_globals, jData, "globals", prefix);
